requirements.txt
pandas
selenium
webdriver_manager
bs4
boto3==1.34.90
botocore==1.34.90
# delta-core
pyspark==3.5.0
apache-airflow-providers-apache-spark==4.9.0
delta-spark
apache-airflow-providers-amazon
streamlit